requirements.txt
flask==3.0.3
twilio==9.3.6
openai==1.51.0
python-dotenv==1.0.1
pyngrok==7.2.0
supabase==2.9.0
requests==2.32.3
